Implement stateful container stop
This makes it possible to have the container save its state at stop
time, then restore its state on restart.

The feature is mostly interesting as a way to do a "suspend to disk"
kinda of equivalent where there is a guarantee that no work will be done
after the tasks are dumped to disk.

Expected use of the feature is to stop containers when more important
containers need the memory resources as well as a way to do a quick host
reboot without loosing running state.

This branch requires the client to specifically ask for state to be save
and restored at both stop and start time. The command line client is set
so that state isn't capture on stop by default (requires --stateful) but
is restored automatically on start (unless --stateless is passed).

Once checkpoint/restore as proved to be reliable, we should probably
introduce a server option, or a container option to have this be used on
host reboot in place of a standard container shutdown.

Closes #1558

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>