Fix DB test
The DB creation script says it can be called even against an existing
DB, but that's simply not true.

With the old code, if a table is added to LXD, the whole new table will
be created ahead of running the migration code which would otherwise
create it, causing errors.

Instead, only apply the schema when no schema is present (new database).

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>