Improve stateful stop
This exports the fact that the container contains stored state through
the same "stateful" property already used by snapshots.

With this information, the client can now only restore a stateful
startup when state is in fact stored.

Which in turn means the daemon can return an error when requested to
restore state for a container which has none.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>