Don't stop at an unsatisfactory sub?id entry
We parse /etc/sub?id and stop when we find an entry for the user.  If
that isn't big enough, we later refuse to start unpriv containers.  But
sub?id may have later entries which are big enough - so just always
make sure to only return entries which will suffice.

Closes #1687

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>