migration: attempt to be slightly smart about moving filesystems
This breaks up live migrations into two parts: first we send a snapshot of
the filesystem while the container is running (and all of its snapshots),
then we do the checkpoint and do a final incremental send. The goal here is
to minimize the amount of downtime for the container.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>