mounting: only block devices hold filesystems
If we lxc config device add x1 disk type=disk source=/dev/null, lxd
should not try to detect the filesystem (and fail when it can't).

The function as it stands woudl've made sense, but every caller of
it really wants to know whether it's a block device, so rename the
fn and have it return false for chardevs.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>