daemon: support proxy configuration options
juju will want to set proxy information on the daemon to tell it where to
ask for images. This commit adds configuration varaibles for common proxy
information, so that juju can use the LXD API as opposed to hacking the
init script to start LXD with HTTP_PROXY env enabled.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>