Make changing https_address more reliable
Rather than try to guess what socket to close and replace through our
list of sockets, just replace the list by two fixed sockets in the
config (unix and tcp) and always close the TCP one.

Also attempt to bind the address again for up to a second in 100ms
increment as the kernel is sometimes a bit slow to release the socket.

Closes #1732

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>