migration: don't defer cleanup of sending snapshots
The problem here is that depending on the migration type, we might delete
the container on the source host directly after the migrate returns, which
can race with defer and cause the delete to fail. Instead, let's explicitly
delete the stuff we need to before returning, so there is no race.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>