Handle unresponsive container monitors
If a container monitor is unresponsive, we may wait forever for
responses over the lxc command socket.

An easy way to reproduce this is to choose a container monitor
process (look for a process like:
	[lxc monitor] /var/lib/lxc/containers containername
) and suspend it with 'kill -STOP'.

So put a one-second timeout around calls to the go-lxc State()
function.  This leads to lxc list feedback like:

0 ✓ serge@sl ~ $ lxc list
+------+-------+------+------+------------+-----------+
| NAME | STATE | IPV4 | IPV6 |    TYPE    | SNAPSHOTS |
+------+-------+------+------+------------+-----------+
| x1   | ERROR |      |      | PERSISTENT | 0         |
+------+-------+------+------+------------+-----------+
0 ✓ serge@sl ~ $ lxc info x1
error: Monitor is hung
1 ✗ serge@sl ~ $ lxc stop x1
error: Monitor is hung

If there were thousands of containers with hung monitors the 1s
each would add up, but this is supposed to be a mitigation for a
rare case.  If we end up with a lot of hung monitors we should
figure out why and prevent it.

Closes #1752

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>