Make it possible to have templates only apply for non-existing files
We unfortunately can't use the suggested "overwrite" keyword because
we'd need it to default to true when unset and the yaml parser doesn't
let us do that.

So instead we call it "create_only" and have it default to false
(default for bool types) which gives us an equivalent behavior.

Closes #1757

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>