exec: wait until the window size is set
While debugging something else, I noticed that the term size for exec is racy:
if the command happens to run before the goroutine in the client sends the
actual size of the terminal, we just get the default size for the pty.

Let's wait until the window size is set at least once before we do exec. It may
be worth documenting that this needs to happen somewhere, but I'm not sure
where :)

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>