rsync: use custom netcat instead of nc -U
For whatever reason, it seemed (through checking via some combination of
strace and printing websocket frame lengths) that netcat was doing some
buffering of the socket contents and causing a hang in some cases. Instead,
let's use our own very dumb implementation of netcat instead, avoiding any
buffering.

Now that we are execing ourself to do rsync, we can't do any tests in go
test, because the test binary is not the lxd binary, and so we'll fork bomb
ourself. So, let's get rid of the rsync test. This doesn't get rid of test
coverage, since we're doing a cold migration in the test suite anyway.

Closes #1944

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>