Implement PKI authentication
This adds support for enteprise PKI mode where a central Certificate
Authority is used to issue certificates for the LXD daemons as well as
the clients.

To enable this, replace or add the following files on the daemon side:
 - server.crt (CA generated server certificate)
 - server.key (key for the above certificate)
 - server.ca (Certificate Authority root certificate)

And on the client side:
 - client.crt (CA generated client certificate)
 - client.key (key for the above certificate)
 - client.ca (Certificate Authority root certificate)

In such mode, only clients using certificates issued by the PKI will be
able to connect. Normal password authentication is still used on first
handshake.

Closes #1985

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>