Add config key for container force shutdown timeout
Currently if a container takes more than 30 seconds to shutdown, it is stopped
anyway. The 30 second timeout is hardcoded and this change makes the
timeout configurable.

This boot.force_shutdown_timeout container config key sets how long to wait
(in seconds) for a container to shutdown before it is force stopped.

This addresses issue https://github.com/lxc/lxd/issues/2027

Signed-off-by: Rene Fragoso <ctrlrsf@gmail.com>