Add push file method that does not send file owner and mode
This method is similar to PushFile, but it does not accept file owner
and mode parameters, and thus does not send any permissions to the lxd
server when pushing the file. By not sending the file permissions to lxd,
lxd will not change the file owner and permissions after file is copied.

This new method is required because PushFile could not be modified
without changing its API to accomplish the same thing. Although uid
and gid could be set to -1, mode is an os.FileMode (uint32) and
setting it to -1 would overflow it.

Signed-off-by: Rene Fragoso <ctrlrsf@gmail.com>