Add support for os.SyscallError in client
Modify the client's error handling to support os.SyscallErrors.  Add
an exported function to the client to handle detection of the special
case errors so that the core logic can be reused by other projects,
e.g. juju.  Generation of the error message is kept out of the client
so as to avoid adding i18n messages into other projects' output.

Update the 'LXD socket not found...' error message to include a prompt
about LXD not being installed.

Add a message to the 'funcs starting with empty line' test to help
users with bad habits understand what they did wrong.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>