Don't show last used at column for remotes running old lxd
Prior versions of lxd will not include a last_used_at field in their
JSON response. When the lxd client unmarshalls the JSON response from
an older remote, the LastUsedDate field will be left at its zero
value. Here we check if the LastUsedDate is this zero value so we don't
show the "last used at" column.

Signed-off-by: Rene Fragoso <ctrlrsf@gmail.com>