Fix a panic in 'remote add'
Use strings.TrimPrefix to strip 'unix://' from the URL/path in both
remoteCmd.addServer() and connectViaUnix.  This resolves the direct
cause of the panic and makes the handling of relative paths consistent
in both flows.

Add a shared util IsUnixSocket() and use it when determining whether
a path without an explicit scheme should be treated as HTTPS or UNIX.

Remove an extra '/' from the path used for TestLocalLXDError, adding
a single slash makes an absolute path look like a relative path with
a 'unix://' scheme.

Fixes #2089

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>