Add target path to rootfs tarball in image export
The split image path didn't pass in 'target' to the Join call.

Add an integration test to verify exporting split images.  Add an
option to the import-busybox script to control whether or not to
set the filename for split images.  Change the effective filename
to be 'busybox', previously it was being set to 'busybox.rootfs'.

Fixes #1980

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>