Always use fingerprint for image export filename
Independently detect the compression algorithm for the metadata and
rootfs files as they may use different compression algorithms, e.g.
XZ for the metadata and squashfs for the rootfs.  Using the original
file name of the metadata as the base file name for all output files
can result an incorrect file extension for the rootfs.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>