Do our own socket activation
go-systemd is great but brings a bunch of dependencies we don't care so
much about when all we want is the tiny bit of code parsing the systemd
environment.

This includes a basic re-implementation of this code done from the
systemd documentation. It basically grabs the two env variables, does a
quick sanity check and opens the fds and turns them into network
listeners.

It also then clears the two env variables so we don't confuse anything else.

Closes #2333

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>