remove logging import
downstream users of shared/ (i.e. juju) don't want to have a log15
dependency.

log.Ctx() is defined as a map[string]interface{}, but the log15 package
does some fancy stuff to normalize it when it gets passed, which is why we
needed to wrap it in a log.Ctx() again. Instead, let's just not unwrap the
type in the first place, which means log15's machinery will see the log.Ctx
users passed us and behave correctly.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>