usb: vendorid is mandatory, productid isn't
The code was doing the right thing, skipping devices that didn't match
vendorid and allowing devices that don't have a productid set to still
be passed through.

But the documentation and configuration parser were doing things the
other way around, requiring a productid and considering vendorid
optional.

So fix the documentation and config parser to do the right thing.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>