disable keepalives in http.Transports
http/transport.go says:

// By default, Transport caches connections for future re-use.
// This may leave many open connections when accessing many hosts.
// This behavior can be managed using Transport's CloseIdleConnections
// method
// and the MaxIdleConnsPerHost and DisableKeepAlives fields.

In particular,

babbageclunk | Yeah - we end up leaking 10 goroutines everytime we destroy a model.

juju uses the client from a long running daemon, which creates a new
http.Transport for each call to the API, which creates a new
http.Transport, which leaks these goroutines.

Perhaps we can fix this more better with the client rewrite, but for now
let's just not do the keepalive, since that's how we're using
http.Transport anyway.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>