lxd/container_lxc, lxd/container: add ExecNoWait()
ExecNoWait() constructs a lxd forkexec command that calls a go-lxc binding which
in turn calls low-level liblxc's c->attach(). In contrast to Exec() ExecNoWait()
calls a go-lxc wraper that does not wait on the command to exit. Rather,
the go-lxc wrapper will simply return the PID it receives from low-level
liblxc's c->attach(). Hence, it returns the PID of the attached process. To
retrieve the attached PID from the go-lxc wrapper ExecNoWait() expects to be
passed the write end of a pipe into which the PID of the executing process will
be written. Callers can then read the PID of the executing process from the read
end of the pipe. Note that the returned PID is a child of the lxd forkexec
command and hence cannot be directly waited upon.

Signed-off-by: Christian Brauner <christian.brauner@canonical.com>