lxd/container_exec: switch to ExecNoWait()
In order to forward signals to the program we are executing we need a function
that allows us to retrieve the PID of the executing process. To this end,
ExecNoWait() epects to be passed the write end of a pipe so that the lxd
forkexec command can write the PID of the executing process into it. The caller
can then read the PID of the executing process from the read end of the pipe.
Note that we cannot directly wait on the PID read from the read end of the pipe
since it is a child of the lxd forkexec command. We can however forward signals
to it.

Signed-off-by: Christian Brauner <christian.brauner@canonical.com>