Store the simplestreams cache to disk
Right now, we store simplestreams data for up to an hour. This is done
through a simple memory cache in LXD.

The goal of this was to decrease load on the simplestreams servers and
that's been working pretty well so far.

One thing that this cache doesn't help much with is dealing with
simplestreams servers being offline or the client being temporarily
disconnected from the internet.

Right now once the cache times out, LXD will refresh it, if that fails,
then the cache is gone.

This commit makes it so that our cache persists across LXD restarts by
serializing it to disk. It also changes its use a bit, so that if we
fail to refresh the cache, we log a warning and bump its expiry for
another hour.

This effectively allows LXD to be used completely offline and work fine
with any image that's already in the local cache.

Closes #2487

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>