Remove legacy code from OnStop
This changes OnStop to assume LXC 2.0.0 or higher.

So it now relies on LXC_TARGET being set in the environment (and will
fail if it's not) and also depends on LXC interrupting container restart
on hook failures.

This allows the removal of a good chunk of code and improved logging of
what's going on with the container.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>