devices: add gpu device type
- lxc config device add c1 gpu0 gpu => Pass everything we have
- lxc config device add c1 gpu0 gpu vendor=10de => Pass whatever we have from nvidia
- lxc config device add c1 gpu0 gpu vendor=10de id=1 => Pass the second nvidia GPU
- lxc config device add c1 gpu0 gpu pci=0000:02:08.0 => Pass whatever GPU is at that PCI address

This parses /sys/bus/pci/devices/* and looks for the drm subfolder. All
subfolders under that, e.g.

            card0
            controlD64
            renderD128

belong to the same PCI address and probably all need to be handed to the
container when the user somehow selects card0.  We then retrieve the vendor ID,
and if available the device ID from

            /sys/bus/pci/devices/*/vendor
            /sys/bus/pci/devices/*/device

The major and minor number for each device can be gathered from the

            card0/dev
            controlD64/dev
            renderD128/dev

files.

If we detect that the vendor ID is Nvidia, we also parse /dev and look for

            /dev/nvidia[0-9]+

entries. We thereby associate the correct /dev/nvidia* entry with the correct
/dev/card* entry, e.g. we correlate

            /dev/card0
    and
            /dev/nvidia0.

Finally, we store any residual /dev/nvidia[^0-9]+ mounts that are not cards
because we need to assume that if /dev/nvidia[0-9]+ is requested then the other
files are necessary for correct functionality of that card and need to be
mounted into every container that also mounts that card.

Signed-off-by: Christian Brauner <christian.brauner@canonical.com>