stateful start: start storage when necessary
Similar to 38d794806a8ca781a0ed69798c8a6e9ea8ea7a70, we need to start the
storage at the right time for stateful start and stop. In particular, the
onstart hook will start the storage for us, but we need to be sure and
start it so that we can uidshift the images.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>