stateful start: start storage when necessary
Similar to 051d733dae9ff396b86a93eacfb9b730b253a17f, we need to start the
storage at the right time for stateful start and stop. In particular, the
onstart hook will start the storage for us, but we need to be sure and
start it so that we can uidshift the images.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>