exec: Exec() return attached PID && take bool arg
The int pointer argument attachedPid was used to communicate to the function
whether the user wants Exec() to run and wait the command or just start to run
it. Let's make it more go-y by giving Exec() an explicit boolean argument "wait"
which the user can use to tell the function whether he wants it to wait on the
command or not and return the attached PID as an additional return value.

Signed-off-by: Christian Brauner <christian.brauner@canonical.com>