implement `lxd import`
The basic idea here is to maintain a file in the container's directory that
indicates what configuration and snapshots it has, so that if someone just
backs up the filesystem the container is on (e.g. the zfs pool), they can
re-import their containers with a simple `lxd import` command.

Closes #2286

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>