Change ContainerStart to take the name and path to start
Instead of taking the container and computing the name and path, let's just
take the name and path themselves. We'll use this in the next patch to
start the storage for a container that we haven't created in memory or in
the database yet, when we're reading its slurp file.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>