file push: fix off by one error
If filepath.Dir(someDir) returns "." we need to substract minus one to include
the first letter of the directory we want to create. This seems to only affect
recursive mode.

Also this adds two tests: One for when we are in the same working directory as
the directory we want to push (This will cover filepath.Dir(someDir) returning
".".). And one for when we give a directory via "../".

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>