shared: add WebsocketExecMirror()
This function specifically deals with shoveling data from an interactive lxc
exec session. It allows to detect background processes. The idea is to detect
when the attached child has exited. We can assume it has only exited after it
was able to write everything it wanted to write into the tty buffer. That means
at maximum output the size of one tty buffer still belongs to the attached
child. Everything afterwards belongs to another (background) process. So we dump
65536 bytes at maximum and then exit.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>