exec: detect background tasks to allow clean exit
This adds the function ExecReaderToChannel() to our linux specific utility
functions. It is the workhorse that reads from the master side of a pty file
descriptor. Background tasks are identified correctly. That means you can run:

	chb@conventiont|~/source/go/bin
	> ./lxc exec xen1 -- bash
	root@xen1:~# sleep infinity &

or

	chb@conventiont|~/source/go/bin
	> ./lxc exec xen1 -- bash
	root@xen1:~# yes &
	[1] 290
	root@xen1:~# y
	root@xen1:~# y
	root@xen1:~# y
	.
	.
	.

and still cleanly exit via "Ctrl+D" or "exit". The function is explained in
detail directly in the code.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>