exec: report exit code when we got killed by signal
The standard shell convention is to report 128 + n where n := signal that killed
the process. This signal number is usually within a reasonable range so that the
magic 8-bit wall of sound is not breached. Essentially, we now report exit codes
like a standard shell would:

1.
	chb@conventiont|~/source/go/bin
	> ./lxc exec zest1 sleep 100
	chb@conventiont|~/source/go/bin
	> echo $?
	143

caused by sending:

	kill -15 $(pidof sleep 100)

2.
	chb@conventiont|~/source/go/bin
	> ./lxc exec zest1 sleep 100
	chb@conventiont|~/source/go/bin
	> echo $?
	129

caused by sending:

	kill -HUP $(pidof sleep 100)

3.
	chb@conventiont|~/source/go/bin
	> ./lxc exec zest1 sleep 100
	chb@conventiont|~/source/go/bin
	> echo $?
	137

caused by sending:

	kill -KILL $(pidof sleep 100)

This is way more fine-grained than what ssh does and more in line with standard
shell exit codes. Also, this allows us to even report back SIGKILL in the sense
that we can tell the user, "Hey, this was the grim 137 aka the ultimate 9.".

Note that there is a caveat: If a user were to write a program that uses one of
those 128 + n exit codes as exit code than being killed by a signal would be
indistinguishable for him from success. However, the user would face the same
problem running the program in a local shell or otherwise. So I wouldn't count
this as a counter-argument. (Also, there is the fact that these exit codes are
actually reserved.)

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>