cloud-init.md: Add Documentation on Network Configuration via cloud-init
At the moment there is no definitive documentation on how to use
cloud-init for custom network configuration of lxd containers. Not to
mention that it is not clear from the current documentation that this
behavior depends on image contents and templates shipped with the image.

This guide is a bit practical and Ubuntu-specific but should give a
good starting point as it describes the implementation used right now
and at least gives a working example and shows the resulting layout.

Signed-off-by: Dmitrii Shcherbakov <dmitrii.shcherbakov@canonical.com>