config: Always use "simplestreams" for images:
We switched to simplestreams by default as the protocol for the images
remote a while back, yet we're still seeing a majority of users use the
LXD protocol.

This will update existing user configuration to use the right protocol
for that remote and hopefully let us one day deprecate the LXD protocol
from images.linuxcontainers.org.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>