container: check whether we can set quota
We call ContainerStorageReady() to ensure that the storage volume for the
container actually exists and if not, we skip ContainerSetQuota(). This e.g.
happens in the create case. But not to worry, the quota will be set on
containerConfigureInternal() which is called after the storage volume for the
container exists.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>