patches: fix zfs upgrade from existing dataset
When we upgrade from an existing dataset we must make sure that we refer to the
right entity in the right places:
- When we want to talk about the pool in the db, we need to use "poolName" which
  will always hold the correct name for the db entry.
- When we want to talk about the pool on-disk, we need to use "defaultPoolName"
  which will always hold the correct on-disk name of the zfs pool.

The good news is that users who are ran this upgrade on an existing dataset will
have not moved any further then the pool-in-db creation step. So simply
replacing their db with the old backed-up one and rerunning the upgrade should
fix all issues since no on-disk information has been touched yet.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>