storage: return correct error
This is more visual cruft but ensures that a non-confusing error is returned.
For example, in a lvm test-run on another branch Jenkins returned:

	INFO[02-27|00:03:30] Creating container                       ephemeral=false name=nonlive
	INFO[02-27|00:03:31] Deleting container                       created=2017-02-27T00:03:30+0000 ephemeral=false name=nonlive used=1970-01-01T00:00:00+0000
	INFO[02-27|00:03:31] Deleted container                        created=2017-02-27T00:03:30+0000 ephemeral=false name=nonlive used=1970-01-01T00:00:00+0000
	EROR[02-27|00:03:31] Failed to initialize container storage   ephemeral=false name=nonlive

	error: Invalid storage type.

As container creation is API internal it is actually not possible that a
non-valid storage driver was passed. So this error stems from a call to
<storage-type>.StoragePoolInit() which was hidden because on any error

	return nil, fmt.Errorf("Invalid storage type.")

was returned. Don't do that!

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>