storage_cgo: make flag argument configurable
For loop-backed btrfs pools we actually mount the loop device and so it's fine
to use LO_FLAGS_AUTOCLEAR since the loop device is guaranteed to be kept around
for as long as the mount exists. When we implement lvm loop-backed pools
however, we do not mount the loop device so we need to make sure that it is not
deconfigured automatically when we close all handles to it.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>