storage: call StoragePoolCheck() if pool exists
The operations performed in StoragePoolCheck() are only operations that are
necessary to guarantee correct functionality of the storage pool. So call them
in storagePoolInit(). To support this, storagePoolInit() gains the boolean
argument "poolExistsOnDisk" which is used to indicate whether the pool has
already been created on disk.
The only exception is in the case of calling StoragePoolCreate(). In this case
StoragePoolCheck() needs to be called explicitly, to make sure that things are
alright after creation.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>