lxc: Rework for better manpages
This commit and the following ones for each individual command are
designed to improve the help messages from the lxc command line tool so
that help2man will generate proper manpages for us.

This includes adding a bunch of spacing to generate proper options,
better marking help sub-sections, ...

It also modifies "lxc manpage" so that instead of spitting out a mostly
useless wall of text, it will instead now generate proper manpages in
the specified directory.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>