patches: delete image db entry if lv is missing
When users mixed DIR and LVM on an pre-storage-api LXD LVM instance some images
might exist in the images database but do not actually exist as logical
volumes. In this case the upgrade code will be confused and error out. If we
detect such a case we should delete the image from the storage volumes
database and move on.

Closes #3026.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>