patches: ensure existing pool config is kept
When the daemon gets killed mid-upgrade and the storage_api patch doesn't
successfully complete but a storage pool entry was already created in the db it
is vital that we do not alter the current config. We should only alter it when
it is not set in the first place.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>