lxd: allow container only migration
This implements copying a container with or without its snapshots by
introducing a "container_only" attribute. This can be set to false or true. For
local container copies this introduces copying the container including its
snapshots and makes this the default. Remote migration requests on the other
hand will now gain the ability to only move the container without its snapshots.

Closes #3006.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>