Honor the LXD_BACKEND environment variable in storage tests
Instead of unconditionally run storage tests depending on the
availability of the underlying backend tools, run them depending on
the value of LXD_BACKEND.

There's a small hack detecting whether we are running as
lxd-github-pull-test Jenkins job, and in that case LXD_BACKEND gets
automatically set to $backend. I think it'd be best to change the
Jenkins configuration instead, and I'm not sure what other jobs would
need a change too (beside lxd-github-pull-test).

Signed-off-by: Free Ekanayaka <free.ekanayaka@gmail.com>