Add a new shared/cmd package with initial command I/O logic
This is a proof of concept of some unit-testable code that I'd like
to later use for making cmdInit testable (as per #2834).

It's small so should it be easy to review and because I wish to have
some feedback on the approach before moving forward. I hope that the
branch gives a feel of what further branches would look like.

As a general note, I feel it'd be important to improve coverage of the
parts that can be unit tested easily (so basically most higher-level
application logic except for system-level interactions). I think this
can be done incrementally without too much disruption, if we agree.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>